Fixed and moved csv headers to config file
For perf stat the column names were wrong. Three empty columns at the end were
missing. This caused pandas to misinterpret the perf stat csv file.
The format used would be fine for older versions of perf.
To make it easy to change the format I moved the column names to defconfig.py -
this is a dirty hack actually since the column names are used at literals
elsewhere. As such they should stay within the code and the user should not
have to figure out the format by himself.