Fixed column formatting to be compatible with latest perf
More changes were required because multiple dataframe columns with same name
will result in bad data types for columns (when selected I got DataFrame instead
of DataSeries). Also, when reading the CSV we have to specify more options to
make sure the csv is correctly parsed.
This is still not cleaned because the column names should be somewhere internal
and not in config but this is now returning proper plots.