Fixed broken streaming plot for multiple events
Streaming perf stat plots were broken. The plot only supported a single event at
a time, showed no legends and silently discarded the rest of the data (it was
still saved to file).
This has been now replaced by a proper plot that supports multiple data sources
that can be updated asynchronously and that are correctly styled and colored.