Added an interval/period input box for perf record.
A simple UI element for choosing the period time for the leader sample (cycles)
is added. The interval time is taken and cpuinfo is used to calculate the number
of events before resampling.
Additionally we check for the nominal and actual frequencies and print a warning
in case of mismatch because we need many of the performance calculations only
work if the processor is actually running at ref-cycle frequency.